+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Improvements
 - Avoid unnecessary panel updates by using an Epsilon comparison when scrolling children
 - Allow setting a default text alignment for paragraphs in UiStyle
 - Made custom values of Element.Style persist when a new ui style is set
+- Update elements less aggressively when changing a ui system's style
 
 Fixes
 - Fixed paragraph links having incorrect hover locations when using special text alignments